migrator: Implement a non-spammy (in errors) move data between tables
This uses an SQL anonymous function to check if the tables exists, before
running the queries on them.

Note that this places some restrictions on the queries, they may not contain
transactions ( because that's not allowed from inside anonymous functions)

However, overall this appears to be a functional way of doing the data migration.