migrator: Use COPY pipeline between databases to migrate data
Doing postgres_fdw data migrations between tables turns out to be utterly slow
and distasteful. This instead wraps it in a pair of COPY operations through a
pipe.

It will cause a temp table to exist for the sorting of the source, which may
inflate disk usage a fair bit, but it means we won't have to worry about insert
order of the archive DB afterwards.