housekeeper: Implement small-batch dedupe function
Deduplication is rather heavy, and slow on BRIN indexed data.  So we should
perform it before we drop BTREE indexes from our tables.

This might mean we should do so while we CLUSTER, or when we perform
maintenance dropping of BTREE indexes.

Since our current DEDUPE is Quadric in runtime and memory, we perform it on a
small amount of data in batches.