Disable "clean expired" and updated it to zabbix4
This updates the "clean expired" code and comments it out.

The code should now be compatible with zabbix4, but it is also disabled
as it is no longer compatible with zabbix3.

Rather than maintain fragile and hard-to-test logic, I'm disabling it as
the cost of it is relatively minor.

Issue: ModioAB/agile#831