Stop using CONCURRENTLY when creating indexes
CONCURRENT indexes are only valid if they are not created inside a
transaction block.

psycopg2 has started to automatically emit BEGIN / END transaction when
using the "with "  context manager, thus causing explicit transactions
to happen for us here.

See: https://github.com/psycopg/psycopg2/issues/941

This means that we now need to stop using concurrently as we may or may
not be inside a transaction block.