archiver: Drop NOT NULL on text archive tables
Zabbix 4.0 has migrated away from "id" columns on text tables as well,
however when we want the archiver to support both old (id ) and new (no
id) we need to alter the archive_text to make them nullable.

The migrator automatically did this change when it ran, but now we need
to make sure it's up to date on the db2 server as well.