Make our own connect conext manager to enable autocommit
With psycopg2 we cannot use the "connect()" as a conext manager anymore
without getting an transaction automatically opened.

This means we need to replace it with our own wrapper function that does
the same, but that also enables autocommit mode for the connection.

This was forced by the upstream behaviour change documented:

>
> Yes, that's the way to do it: don't use a with block if you don't want
> a transaction to start.

https://github.com/psycopg/psycopg2/issues/941#issuecomment-864025101