Archiver: Rollback connection if tables are already swapped
The changes in psycopg2 keeps hitting us. Here, we need to rollback the
connection if we fail to swap tables, a case that happens if the
housekeeper has failed during previous run (due to interruption, or
similar)

In that case, we already have swapped the live & archive table, but not
transferred the data, therefore we cannot do the swap again.

However, that failure is a good case for us, so we need to roll back the
transaction and start again.